keda-azure-storage-queue/requirements.txt
azure-core==1.24.2
azure-identity==1.10.0
azure-storage-queue==12.4.0
certifi==2022.6.15
cffi==1.15.1
charset-normalizer==2.1.0
click==8.1.3
colorama==0.4.5
cryptography==37.0.4
Flask==2.1.3
gunicorn==20.1.0
idna==3.3
importlib-metadata==4.12.0
isodate==0.6.1
itsdangerous==2.1.2
Jinja2==3.1.2
MarkupSafe==2.1.1
msal==1.18.0
msal-extensions==1.0.0
msrest==0.7.1
oauthlib==3.2.0
portalocker==2.5.1
pycparser==2.21
PyJWT==2.4.0
python-dotenv==0.20.0
requests==2.28.1
requests-oauthlib==1.3.1
six==1.16.0
typing-extensions==4.3.0
urllib3==1.26.11
Werkzeug==2.2.1
zipp==3.8.1
